Moved permissions to a separate debug submenu
